refactor: move pings.ts
