+++ packages/dht/src/exports.ts
@@ -2,6 +2,7 @@ export { DhtNode, DhtNodeEvents, DhtNodeOptions } from './dht/DhtNode'
 export { ListeningRpcCommunicator } from './transport/ListeningRpcCommunicator'
 export { Simulator, LatencyType } from './connection/Simulator/Simulator'
 export { SimulatorTransport } from './connection/Simulator/SimulatorTransport'
+export { getRandomRegion, getRegionDelayMatrix } from './connection/Simulator/pings'
 export { PeerDescriptor, Message, NodeType } from './proto/packages/dht/protos/DhtRpc'
 export { ITransport } from './transport/ITransport'
 export { ConnectionManager, ConnectionLocker } from './connection/ConnectionManager'
