+++ packages/dht/test/utils/utils.ts
@@ -27,7 +27,7 @@ import { ServerCallContext } from '@protobuf-ts/runtime-rpc'
 import { Simulator } from '../../src/connection/Simulator/Simulator'
 import { ConnectionManager } from '../../src/connection/ConnectionManager'
 import { v4 } from 'uuid'
-import { getRandomRegion } from '../data/pings'
+import { getRandomRegion } from '../../src/connection/Simulator/pings'
 import { Empty } from '../../src/proto/google/protobuf/empty'
 import { Any } from '../../src/proto/google/protobuf/any'
 import { waitForCondition } from '@streamr/utils'
